requirements.txt
streamlit
langchain==0.1.13
docx2txt
langchain_google_genai
# chromadb
transformers 
optimum 
accelerate
faiss-cpu
openai-whisper
langchain-community